ERAv3/ERAV3_S3_Assignment1_v1/requirements.txt
flask
torch
torchvision
torchaudio
pillow
soundfile
trimesh
numpy
nltk
textblob
matplotlib
pytorch3d